# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Notbook Study API",
    "description": "S0 local vault: notebooks, ingest (PDF / markdown / paste / image OCR), inspectable chunks, retrieve with citation ids, and an inference adapter.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inference": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Inference Info",
        "operationId": "inference_info_api_v1_inference_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceInfoOut"
                }
              }
            }
          }
        }
      }
    },
    "/inference": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Inference Info",
        "operationId": "inference_info_inference_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceInfoOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks": {
      "get": {
        "tags": [
          "notebooks"
        ],
        "summary": "List Notebooks",
        "operationId": "list_notebooks_api_v1_notebooks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotebookOut"
                  },
                  "type": "array",
                  "title": "Response List Notebooks Api V1 Notebooks Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notebooks"
        ],
        "summary": "Create Notebook",
        "operationId": "create_notebook_api_v1_notebooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}": {
      "get": {
        "tags": [
          "notebooks"
        ],
        "summary": "Get Notebook Route",
        "operationId": "get_notebook_route_api_v1_notebooks__notebook_id__get",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}/sources/upload": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Upload Source",
        "operationId": "upload_source_api_v1_notebooks__notebook_id__sources_upload_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_api_v1_notebooks__notebook_id__sources_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}/sources/paste": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Paste Source",
        "operationId": "paste_source_api_v1_notebooks__notebook_id__sources_paste_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}/retrieve": {
      "post": {
        "tags": [
          "retrieve"
        ],
        "summary": "Retrieve Chunks",
        "operationId": "retrieve_chunks_api_v1_notebooks__notebook_id__retrieve_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}/sources": {
      "get": {
        "tags": [
          "inspect"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_notebooks__notebook_id__sources_get",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceOut"
                  },
                  "title": "Response List Sources Api V1 Notebooks  Notebook Id  Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notebooks/{notebook_id}/sources/{source_id}/chunks": {
      "get": {
        "tags": [
          "inspect"
        ],
        "summary": "List Source Chunks",
        "operationId": "list_source_chunks_api_v1_notebooks__notebook_id__sources__source_id__chunks_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChunkPreview"
                  },
                  "title": "Response List Source Chunks Api V1 Notebooks  Notebook Id  Sources  Source Id  Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chunks/{chunk_id}": {
      "get": {
        "tags": [
          "inspect"
        ],
        "summary": "Get Chunk",
        "operationId": "get_chunk_api_v1_chunks__chunk_id__get",
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_source_api_v1_notebooks__notebook_id__sources_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_source_api_v1_notebooks__notebook_id__sources_upload_post"
      },
      "ChunkDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "notebook_id": {
            "type": "string",
            "title": "Notebook Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Locator"
          },
          "source": {
            "$ref": "#/components/schemas/SourceMeta"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "notebook_id",
          "text",
          "locator",
          "source"
        ],
        "title": "ChunkDetail"
      },
      "ChunkPreview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Locator"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "text_preview": {
            "type": "string",
            "title": "Text Preview"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "locator",
          "text",
          "text_preview"
        ],
        "title": "ChunkPreview"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "service": {
            "type": "string",
            "title": "Service",
            "default": "notbook-study-api"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthOut"
      },
      "InferenceInfoOut": {
        "properties": {
          "adapter": {
            "type": "string",
            "title": "Adapter"
          },
          "embed_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Model"
          },
          "chat_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Model"
          }
        },
        "type": "object",
        "required": [
          "adapter"
        ],
        "title": "InferenceInfoOut"
      },
      "NotebookCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "NotebookCreate"
      },
      "NotebookOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "NotebookOut"
      },
      "PasteIn": {
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PasteIn"
      },
      "RetrieveChunk": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Locator"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "text",
          "locator",
          "score"
        ],
        "title": "RetrieveChunk"
      },
      "RetrieveIn": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveIn"
      },
      "RetrieveOut": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/RetrieveChunk"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "chunks"
        ],
        "title": "RetrieveOut"
      },
      "SourceMeta": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "extract_status": {
            "type": "string",
            "title": "Extract Status"
          },
          "notebook_id": {
            "type": "string",
            "title": "Notebook Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "type",
          "extract_status",
          "notebook_id"
        ],
        "title": "SourceMeta"
      },
      "SourceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "notebook_id": {
            "type": "string",
            "title": "Notebook Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "raw_path": {
            "type": "string",
            "title": "Raw Path"
          },
          "extract_status": {
            "type": "string",
            "title": "Extract Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "notebook_id",
          "filename",
          "type",
          "raw_path",
          "extract_status",
          "created_at"
        ],
        "title": "SourceOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}